# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SwarmUI API
  description: |
    API documentation for SwarmUI - AI-powered development environment with parallel CLI agent orchestration.
    
    ## Overview
    SwarmUI provides a comprehensive API for managing:
    - Chat sessions with AI agents
    - Project and ticket management
    - Background job processing
    - Git operations
    - File system operations
    - Terminal sessions
    - Testing and debugging
    
    ## Authentication
    Most endpoints require authentication via NextAuth.js session cookies.
    
    ## WebSocket
    Real-time updates are available via WebSocket at `ws://localhost:3000`.
    See [WebSocket API documentation](./WEBSOCKET_API.md) for details.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: SwarmUI Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:4000
    description: Alternative dev port
...
paths:
...
  /api/git/branches:
    get:
      tags: [Git]
      summary: List branches
      operationId: listBranches
      responses:
        '200':
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GitBranch'
    post:
      tags: [Git]
      summary: Create branch
      operationId: createBranch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                startPoint:
                  type: string
      responses:
        '201':
          description: Branch created
...
  /api/files/{path}:
    get:
      tags: [Files]
      summary: Read file content
      operationId: readFile
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
          description: File path (URL encoded)
      responses:
        '200':
          description: File content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                  path:
                    type: string
    put:
      tags: [Files]
      summary: Write file content
      operationId: writeFile
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        '200':
          description: File written
    delete:
      tags: [Files]
      summary: Delete file
      operationId: deleteFile
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File deleted
  /api/files/search:
    get:
      tags: [Files]
      summary: Search files
      operationId: searchFiles
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: path
          in: query
          schema:
            type: string
          description: Base path to search in
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileSearchResult'
...
  /api/extensions/{id}:
    get:
      tags: [Extensions]
      summary: Get extension
      operationId: getExtension
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Extension details
    delete:
      tags: [Extensions]
      summary: Uninstall extension
      operationId: uninstallExtension
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Extension uninstalled
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List users
      operationId: listUsers
      security:
        - sessionAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - requires admin role
    post:
      tags: [Admin]
      summary: Create user
      operationId: createUser
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                role:
                  $ref: '#/components/schemas/UserRole'
      responses:
        '201':
          description: User created
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    GitBranch:
      type: object
      properties:
        name:
          type: string
        upstream:
          type: string
          nullable: true
        current:
          type: boolean
        isRemote:
          type: boolean
...
    FileSearchResult:
      type: object
      properties:
        path:
          type: string
        name:
          type: string
        matches:
          type: array
          items:
            type: object
            properties:
              line:
                type: integer
              content:
                type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: integer
        updatedAt:
          type: integer
    UserRole:
      type: string
      enum: [admin, editor, viewer]
...